Create Api sub module within CommonPlatform module
