+++ app/controllers/api/internal/v1/hearings_controller.rb
@@ -5,7 +5,7 @@ module Api
     module V1
       class HearingsController < ApplicationController
         def show
-          @hearing = CommonPlatformApi::GetHearingResults.call(hearing_id: params[:id])
+          @hearing = CommonPlatform::Api::GetHearingResults.call(hearing_id: params[:id])
           render json: HearingSerializer.new(@hearing, serialization_options)
         end
