+++ app/controllers/api/internal/v2/hearings_controller.rb
@@ -5,7 +5,7 @@ module Api
     module V2
       class HearingsController < ApplicationController
         def show
-          @hearing = CommonPlatformApi::GetHearingResults.call(hearing_id: params[:id])
+          @hearing = CommonPlatform::Api::GetHearingResults.call(hearing_id: params[:id])
           render json: HearingSerializer.new(@hearing, serialization_options)
         end
