+++ app/services/common_platform/api/get_hearing_events.rb
@@ -1,0 +1,25 @@
+# frozen_string_literal: true
+
+module CommonPlatform
+  module Api
+    class GetHearingEvents < ApplicationService
+      def initialize(hearing_id:, hearing_date:)
+        @hearing_id = hearing_id
+        @hearing_date = hearing_date
+        @response = HearingEventsFetcher.call(hearing_id: hearing_id, hearing_date: hearing_date)
+      end
+
+      def call
+        HearingEventsRecorder.call(hearing_id: hearing_id, hearing_date: hearing_date, body: response.body) if successful_response?
+      end
+
+    private
+
+      def successful_response?
+        response.status == 200 && response.body.present?
+      end
+
+      attr_reader :hearing_id, :hearing_date, :response
+    end
+  end
+end
