+++ app/services/maat_link_creator.rb
@@ -39,7 +39,7 @@ private
   end
 
   def post_laa_reference_to_common_platform(offence)
-    CommonPlatformApi::RecordLaaReference.call(
+    CommonPlatform::Api::RecordLaaReference.call(
       prosecution_case_id: offence.prosecution_case_id,
       defendant_id: offence.defendant_id,
       offence_id: offence.offence_id,
