+++ app/workers/past_hearings_fetcher_worker.rb
@@ -5,7 +5,7 @@ class PastHearingsFetcherWorker
 
   def perform(request_id, prosecution_case_id)
     Current.set(request_id: request_id) do
-      CommonPlatformApi::ProsecutionCaseHearingsFetcher.call(prosecution_case_id: prosecution_case_id)
+      CommonPlatform::Api::ProsecutionCaseHearingsFetcher.call(prosecution_case_id: prosecution_case_id)
     end
   end
 end
