+++ spec/models/hearing_spec.rb
@@ -16,7 +16,7 @@ RSpec.describe Hearing, type: :model do
       let(:hearing_id) { "4d01840d-5959-4539-a450-d39f57171036" }
       let(:hearing) do
         VCR.use_cassette("hearing_result_fetcher/success") do
-          CommonPlatformApi::GetHearingResults.call(hearing_id: hearing_id)
+          CommonPlatform::Api::GetHearingResults.call(hearing_id: hearing_id)
         end
       end
 
@@ -38,14 +38,14 @@ RSpec.describe Hearing, type: :model do
 
         let(:hearing_event_recording) do
           VCR.use_cassette("hearing_logs_fetcher/success") do
-            CommonPlatformApi::GetHearingEvents.call(hearing_id: hearing_id, hearing_date: hearing_day)
+            CommonPlatform::Api::GetHearingEvents.call(hearing_id: hearing_id, hearing_date: hearing_day)
           end
         end
 
         let(:hearing_events) { [hearing_event_recording] }
 
         before do
-          allow(CommonPlatformApi::GetHearingEvents).to receive(:call).and_return(hearing_events)
+          allow(CommonPlatform::Api::GetHearingEvents).to receive(:call).and_return(hearing_events)
         end
 
         it { expect(hearing.hearing_events).to all be_a(HearingEvent) }
@@ -69,7 +69,7 @@ RSpec.describe Hearing, type: :model do
       let(:hearing_id) { "29b73d8f-7683-4e27-9069-f7a031672c35" }
       let(:hearing) do
         VCR.use_cassette("hearing_result_fetcher/success_hearing_attendees") do
-          CommonPlatformApi::GetHearingResults.call(hearing_id: hearing_id)
+          CommonPlatform::Api::GetHearingResults.call(hearing_id: hearing_id)
         end
       end
 
@@ -101,7 +101,7 @@ RSpec.describe Hearing, type: :model do
       let(:hearing_id) { "da124701-048f-408c-85b4-81138316ddce" }
       let(:hearing) do
         VCR.use_cassette("hearing_result_fetcher/success_hearing_cracked_trial") do
-          CommonPlatformApi::GetHearingResults.call(hearing_id: hearing_id)
+          CommonPlatform::Api::GetHearingResults.call(hearing_id: hearing_id)
         end
       end
