+++ spec/services/common_platform/api/get_hearing_results_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe CommonPlatformApi::GetHearingResults do
+RSpec.describe CommonPlatform::Api::GetHearingResults do
   subject(:get_hearing_results) { described_class.call(hearing_id: hearing_id) }
 
   let(:hearing_id) { "ceb158e3-7171-40ce-915b-441e2c4e3f75" }
@@ -8,7 +8,7 @@ RSpec.describe CommonPlatformApi::GetHearingResults do
   let(:response) { double(body: { amazing_body: true }, status: 200) }
 
   before do
-    allow(CommonPlatformApi::HearingFetcher).to receive(:call).with(hearing_id: hearing_id).and_return(response)
+    allow(CommonPlatform::Api::HearingFetcher).to receive(:call).with(hearing_id: hearing_id).and_return(response)
   end
 
   it "calls the HearingRecorder service" do
