+++ spec/services/common_platform/api/record_representation_order_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe CommonPlatformApi::RecordRepresentationOrder do
+RSpec.describe CommonPlatform::Api::RecordRepresentationOrder do
   subject(:record_representation_order) { described_class.call(params) }
 
   let(:prosecution_case) { ProsecutionCase.create!(id: "5edd67eb-9d8c-44f2-a57e-c8d026defaa4", body: "{}") }
@@ -43,7 +43,7 @@ RSpec.describe CommonPlatformApi::RecordRepresentationOrder do
   end
 
   context "with connection" do
-    let(:connection) { double("CommonPlatformApi::Connection") }
+    let(:connection) { double("CommonPlatform::Connection") }
     let(:request_params) do
       {
         statusCode: "ABCDEF",
